services/log-aggregator/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# MongoDB
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
schedule==1.2.0

# For development
pytest==7.4.3
httpx==0.25.2
